* Add using namespace mv

* Resolve type warnings

* Resolve type warning

* Fix includes

---------

+++ src/GeneSurferPlugin.cpp
@@ -5,6 +5,7 @@
 
 #include <util/Serialization.h>
 #include <DatasetsMimeData.h>
+#include <util/Serialization.h>
 
 #include <vector>
 #include <random>
@@ -33,7 +34,6 @@
 #include <string>
 
 #include <chrono>
-
 
 
 Q_PLUGIN_METADATA(IID "nl.BioVault.GeneSurferPlugin")
@@ -1472,7 +1472,7 @@ void GeneSurferPlugin::updateSingleCellOption() {
         _enabledDimNames = _geneNamesAvgExpr;
 
         // update max number of genes in _numGenesThresholdAction
-        _settingsAction.getClusteringAction().getNumGenesThresholdAction().setMaximum(_enabledDimNames.size());
+        _settingsAction.getClusteringAction().getNumGenesThresholdAction().setMaximum(static_cast<int>(_enabledDimNames.size()));
 
         updateSelection();
     }
@@ -1491,7 +1491,7 @@ void GeneSurferPlugin::updateSingleCellOption() {
         }
 
         // update max number of genes in _numGenesThresholdAction
-        _settingsAction.getClusteringAction().getNumGenesThresholdAction().setMaximum(_enabledDimNames.size());
+        _settingsAction.getClusteringAction().getNumGenesThresholdAction().setMaximum(static_cast<int>(_enabledDimNames.size()));
 
         updateSelection();
     }
@@ -1750,7 +1750,7 @@ void GeneSurferPlugin::updateRNAData()
     //qDebug() << "_enabledDimNames size: " << _enabledDimNames.size() << _enabledDimNames[0] << _enabledDimNames[1] << _enabledDimNames[2];
 
     // update max number of genes in _numGenesThresholdAction
-    _settingsAction.getClusteringAction().getNumGenesThresholdAction().setMaximum(_enabledDimNames.size());
+    _settingsAction.getClusteringAction().getNumGenesThresholdAction().setMaximum(static_cast<int>(_enabledDimNames.size()));
 
     updateSelection();
 }
@@ -2043,8 +2043,8 @@ void GeneSurferPlugin::loadAvgExpressionFromFile() {
         matrixData.push_back(rowData);
     }
 
-    int numClusters = _clusterNamesAvgExpr.size();
-    int numGenes = _geneNamesAvgExpr.size();
+    int numClusters = static_cast<int>(_clusterNamesAvgExpr.size());
+    int numGenes = static_cast<int>(_geneNamesAvgExpr.size());
 
     file.close();
